Fuites de mémoire enlevées + sprites pour maps (mur)

+++ kill.sh
@@ -1,3 +1,7 @@
 #!/bin/bash
-
-kill -9 $(ps aux | grep $1 | tr -s ' ' | cut -f2 -d' ' | cut -f1 -d$'\n')
+if [[ $1 != "" ]]
+then
+	kill -9 $(ps aux | grep $1 | tr -s ' ' | cut -f2 -d' ' | cut -f1 -d$'\n')
+else
+	echo "Usage : ./kill.sh prog_name"
+fi
